+++ client/src/pages/teachers/home.css
@@ -1,0 +1,219 @@
+/* home.css */
+
+/* Main Content Styles */
+.main-content {
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 0 1rem;
+  box-sizing: border-box;
+}
+
+.dashboard-header {
+  text-align: center;
+  margin-bottom: 2rem;
+}
+
+.dashboard-title {
+  font-size: 2.5rem;
+  color: #64ffda; /* Teal accent */
+  text-shadow: 0 0 8px #64ffda;
+}
+
+.dashboard-subtitle {
+  font-size: 1.2rem;
+  color: #ccd6f6; /* Light blue-gray */
+}
+
+.error-message {
+  background: #ff6b6b; /* Coral red for errors */
+  color: white;
+  padding: 1rem;
+  border-radius: 8px;
+  margin: 1rem 0;
+  text-align: center;
+}
+
+/* Card Styles */
+.card-neon {
+  background: #112240; /* Slightly lighter dark blue */
+  border: 2px solid #64ffda; /* Teal border */
+  box-shadow: 0 0 12px #64ffda; /* Teal glow */
+  border-radius: 12px;
+  padding: 1.5rem;
+  margin-bottom: 2rem;
+}
+
+/* Upload Section Styles */
+.upload-section h2 {
+  font-size: 1.8rem;
+  color: #64ffda; /* Teal accent */
+  margin-bottom: 1rem;
+}
+
+.upload-container {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.file-upload-label {
+  display: inline-block;
+  cursor: pointer;
+}
+
+.upload-button {
+  background: #64ffda; /* Teal for buttons */
+  color: #0a192f; /* Dark blue text */
+  padding: 0.75rem 1.5rem;
+  border-radius: 8px;
+  transition: background 0.3s ease;
+}
+
+.upload-button:hover {
+  background: #52d1b2; /* Darker teal on hover */
+}
+
+.file-input {
+  display: none;
+}
+
+.cropper-container {
+  width: 100%;
+  max-width: 800px; /* Fixed maximum width */
+  margin: 0 auto;
+  background: #112240;
+  padding: 1rem;
+  border-radius: 12px;
+  border: 2px solid #64ffda;
+}
+
+.region-selection-controls {
+  display: flex;
+  gap: 1rem;
+  margin-top: 1rem;
+  justify-content: center;
+}
+
+.select-region-button {
+  margin-top: 10px;
+  padding: 10px 20px;
+  background: #64ffda; /* Teal for buttons */
+  color: #0a192f; /* Dark blue text */
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: background 0.3s ease;
+}
+
+.select-region-button:hover {
+  background: #52d1b2; /* Darker teal on hover */
+}
+
+.select-another-button {
+  background: #64ffda;
+  color: #0a192f;
+  padding: 0.75rem 1.5rem;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: background 0.3s ease;
+}
+
+.done-selecting-button {
+  background: #ff6b6b;
+  color: white;
+  padding: 0.75rem 1.5rem;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: background 0.3s ease;
+}
+
+.notes-addition-container {
+  margin-top: 2rem;
+  padding: 1rem;
+  background: #112240;
+  border-radius: 12px;
+  border: 2px solid #64ffda;
+}
+
+.original-image-container {
+  position: relative;
+  width: 800px;
+  height: 600px;
+  margin: 0 auto;
+  overflow: hidden;
+}
+
+.base-image {
+  width: 100%;
+  height: 100%;
+  object-fit: contain;
+  display: block;
+}
+
+.region-note-section {
+  position: relative;
+  cursor: pointer;
+}
+
+.region-outline {
+  transition: filter 0.3s ease;
+}
+
+.note-input-popup {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background: #1a1f2e;
+  padding: 2rem;
+  border-radius: 12px;
+  border: 2px solid #64ffda;
+  z-index: 1000;
+  width: 90%;
+  max-width: 500px;
+}
+
+.notes-textarea {
+  width: 100%;
+  height: 100px;
+  padding: 0.75rem;
+  border: 1px solid #64ffda; /* Teal border */
+  border-radius: 8px;
+  background: rgba(255, 255, 255, 0.1); /* Semi-transparent white */
+  color: white;
+  font-size: 1rem;
+  resize: none;
+  margin-bottom: 1rem;
+}
+
+.save-note-button {
+  margin-top: 20px;
+  padding: 10px 20px;
+  background: #64ffda; /* Teal for buttons */
+  color: #0a192f; /* Dark blue text */
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: background 0.3s ease;
+}
+
+.save-note-button:hover {
+  background: #52d1b2; /* Darker teal on hover */
+}
+
+.done-notes-button {
+  margin: 2rem auto;
+  display: block;
+  background: #64ffda;
+  color: #0a192f;
+  padding: 0.75rem 1.5rem;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+}
+
+.final-output-container {
+  /* Add styles if you expand this section later */
+}
